requirements.txt
pdfplumber
BeautifulSoup4
Pillow
pytesseract
pandas
